FIX 2: No se estaban guardando los cambios en company y rfc en nuevos usuarios desde el panel de control debido a que se debían primero que crear los morph keys, lo cual ya se arregla y se valida a través de una variable en el Eloquen Model que permite solo esos atributos a ser creados en caso de que no existan. FIX 2: OpenpayService mandaba error porque si no se definia la dirección completa con city, state y demás, el API nos devolvia error y una exception, lo que se opto por hacer es una validación antes de intentar hacer el envío de data al API y si falla no procedemos con la actualización.

+++ app/Services/OpenpayService.php
@@ -8,6 +8,7 @@ use App\Exceptions\OpenpayCustomerNotFound;
 use App\User;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Validator;
 use Openpay;
 use OpenpayApiRequestError;
 use Wave\OpenpayUser;
@@ -21,6 +22,13 @@ class OpenpayService
     private $merchant_id;
     private $public_key;
     private $private_key;
+
+    private $address_rules = [
+        'postal_code' => 'required|string|min:1|max:12',
+        'city' => 'required',
+        'line1' => 'required',
+        'state' => 'required',
+    ];
 
     /**
      * Crea el objeto pasando las llaves asignadas en el portal
@@ -87,6 +95,14 @@ class OpenpayService
         // Debido a que no hay forma de actualizar la dirección  directamente al cliente
         // lo que haremos será borrar el cliente y volverlo a crear pero con la dirección actualizada.
         if ($data['type'] != 'billing') return false;
+
+        /**
+         * address.postal_code la longitud tiene que estar entre 1 y 12, address.postal_code is required, address.city
+         * is required, address.line1 is required, address.state is required
+         *
+         * If fails the validation let dont to anything.
+         */
+        if(Validator::make($data,$this->address_rules)->fails()) return false;
         $this->deleteCustomer($user);
         $user->openpay()->delete();
         return $this->createCustomer($user);
